script update

+++ copyTemplate.js
@@ -9,7 +9,13 @@ const destinationDir = process.cwd();
 
 async function copyTemplate() {
   try {
-    console.log('Copying template files...');
+    console.log('Checking template directory...');
+    if (!await fs.pathExists(templateDir)) {
+      console.error(`Template directory does not exist: ${templateDir}`);
+      process.exit(1);
+    }
+
+    console.log('Copying template files from', templateDir, 'to', destinationDir);
     await fs.copy(templateDir, destinationDir);
     console.log('Template copied successfully!');
     installDependencies();
@@ -20,7 +26,7 @@ async function copyTemplate() {
 
 function installDependencies() {
   console.log('Starting npm install...');
-  const install = spawn('npm', ['install'], { stdio: 'inherit', shell: true });
+  const install = spawn('npm', ['install'], { stdio: 'inherit', shell: true, cwd: destinationDir });
 
   install.on('error', (err) => {
     console.error('Failed to start npm install:', err);
